Updated HRM model, began working on up-pass down-pass hrm model
